[java] Split TestKuduClient into smaller modules
The monolithic TestKuduClient class (~2600 lines with 50 tests) was
causing regular timeouts on TSAN builds. This commit splits it into
6 focused test modules that can run independently, significantly
reducing timeout risk on slow builds.

Test modules:
- TestKuduClientBasic (7 tests): Basic table operations, timestamps,
  and soft delete functionality
- TestKuduClientDataTypes (8 tests): Data type tests including strings,
  VARCHAR, binary, timestamps, dates, and decimals
- TestKuduClientScanner (7 tests): Scanner lifecycle, expiration,
  keep-alive, limits, predicates, and non-covered ranges
- TestKuduClientAdvanced (8 tests): Auto-incrementing columns, array
  types, and custom hash schemas per range partition
- TestKuduClientConcurrency (10 tests): Concurrent operations,
  read-your-writes consistency, concurrent flush, and session lifecycle
- TestKuduClientSecurity (10 tests): Authentication, certificates,
  location assignment, and cluster ID

The original TestKuduClient.java has been removed entirely. CI/Jenkins
should run the individual test classes rather than a suite wrapper.

Verified with dist-test: 1000 runs of each individual module complete
successfully without timeouts.

Reviewed-on: http://gerrit.cloudera.org:8080/23609
Tested-by: Kudu Jenkins
